App cleaning - remove unused files
