+++ .idea/misc.xml
@@ -6,6 +6,7 @@
         <entry key="..\:/EICNAM/FIP2/Android/SpaceAppComplete/SpaceApp/app/src/main/res/layout/activity_main.xml" value="0.25416666666666665" />
         <entry key="..\:/EICNAM/FIP2/Android/SpaceAppComplete/SpaceApp/app/src/main/res/layout/card_upcoming_events.xml" value="0.25416666666666665" />
         <entry key="..\:/EICNAM/FIP2/Android/SpaceAppComplete/SpaceApp/app/src/main/res/layout/fragment_home.xml" value="0.25416666666666665" />
+        <entry key="..\:/EICNAM/FIP2/Android/SpaceAppComplete/SpaceApp/app/src/main/res/layout/fragment_iss.xml" value="0.25416666666666665" />
         <entry key="..\:/EICNAM/FIP2/Android/SpaceAppComplete/SpaceApp/app/src/main/res/layout/fragment_upcoming_events.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/noell/AndroidStudioProjects/FIP2/SpaceApp/app/src/main/res/layout/activity_main.xml" value="0.16055625790139064" />
         <entry key="..\:/Users/noell/AndroidStudioProjects/FIP2/SpaceApp/app/src/main/res/layout/fragment_favorite.xml" value="0.1796875" />
