--- app/src/main/java/com/example/spaceapp/ui/upcomingEvt/UpcomingEvents.kt
@@ -1,44 +1,0 @@
-package com.example.spaceapp.ui.upcomingEvt
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
-import com.example.spaceapp.R
-import com.example.spaceapp.RecyclerAdapter
-import res.layout.*
-
-
-/**
- * A simple [Fragment] subclass.
- */
-class UpcomingEvents : Fragment() {
-
-
-    // --------------------------- Recycler View display ------------- //
-
-    private var layoutManager : RecyclerView.LayoutManager? = null
-    private var adapter : RecyclerView.Adapter<RecyclerAdapter.ViewHolder>? = null
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_upcoming_events, container, false)
-    }
-
-    override fun onViewCreated(itemView: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(itemView, savedInstanceState)
-        R.id.recycler_view.apply {
-            // set a LinearLayoutManager to handle Android
-            // RecyclerView behavior
-            layoutManager = LinearLayoutManager(activity)
-            // set the custom adapter to the RecyclerView
-            adapter = RecyclerAdapter()
-        }
-    }
-}
